renamed and moved TopNav and added param for adding sidebar clas
